coprocessor, util: reorganize xeval, codec dirs
