+++ src/coprocessor/codec/mysql/mod.rs
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use util::codec::Result;
+use super::Result;
 use util::escape;
 
 /// `UN_SPECIFIED_FSP` is the unspecified fractional seconds part.
